frontend/requirements.txt
Flask
requests
numpy
python-dotenv